Improved UI for front end using search box for replication groups.
Kept accordion UI as a config option; it's okay for small lists of replications
but loses a lot of useability for large lists.  Killed the template fragment
in lieu of JSON.  (TODO: factor out the common elements of the two templates
to a base, get rid of the rest of the template fragments where feasible.)

Also cleaned up query of individual syncs.