Refactored templates for web service.
Made summaries of replications and connectors Ajax calls that return JSON,
allowing us to eliminate HTML fragments and offload more of the rendering
to the browser client.  Cleaned up the tree of replication states we've been
passing into a simple green/yellow/red dictionary.  Factored a base template
out of the two options for the front page index.